// src/types/org.ts
export interface Org {
    id: string
    username: string 
    nickname: string 

    avatar?: string 
    cover?: string 

    email?: string 
    website?: string 
    location?: string 
    tagline?: string 
    about?: string 
    stack?: string 

    twitter?: string 
    github?: string 
    weibo?: string 

    created?: string
    role?: string
}